@@ -35,31 +35,28 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.6
-  firebase_core: ^3.2.0
-  country_picker: ^2.0.26
-  firebase_auth: ^5.1.2
-  cloud_firestore: ^5.1.0
-  flutter_riverpod:
-  firebase_storage:
-  image_picker:
-  flutter_contacts:
+  firebase_core: ^4.4.0
+  country_picker: ^2.0.27
+  firebase_auth: ^6.1.4
+  cloud_firestore: ^6.1.2
+  flutter_riverpod: ^3.2.0
+  firebase_storage: ^13.0.6
+  image_picker: ^1.1.2
+  flutter_contacts: ^1.1.9+2
   uuid: ^4.4.2
-  intl: ^0.19.0
-  cached_network_image: ^3.4.0
-  cached_video_player: ^2.0.4
-  permission_handler: ^11.0.0
-  emoji_picker_flutter: ^2.2.0
-  enough_giphy_flutter: ^0.4.1
-  enough_platform_widgets: ^0.7.4
-  flutter_platform_widgets: ^3.3.5
-  flutter_sound: ^9.9.5
-  path_provider: ^2.1.4
-  audioplayers: ^6.0.0
+  intl: 0.20.2
+  cached_network_image: ^3.4.1
+  permission_handler: ^11.3.1
+  emoji_picker_flutter: ^4.4.0
+  giphy_get: ^3.6.0
+  flutter_sound: ^9.28.0
+  path_provider: ^2.1.5
+  audioplayers: ^6.4.0
   swipe_to: ^1.0.6
-  firebase_app_check: ^0.3.0+4
-  story_view: ^0.16.0
-  flutter_lints: ^3.0.0
-  agora_uikit: 1.3.10
+  firebase_app_check: ^0.4.1+4
+  story_view: ^0.16.5
+  flutter_lints: ^3.0.2
+  # agora_uikit: ^1.3.10  # TODO: Upgrade when compatible with Flutter 3.38
 
 
 
@@ -75,7 +72,6 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^3.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
