@@ -12,7 +12,7 @@ final groupControllerProvider = Provider((ref) {
 
 class GroupController {
   final GroupRepository groupRepository;
-  final ProviderRef ref;
+  final Ref ref;
 
   GroupController({required this.groupRepository, required this.ref});
   void createGroup(BuildContext context, String name, File profilePic,
