@@ -14,42 +14,45 @@ class LandingScreen extends StatelessWidget {
     final size = MediaQuery.of(context).size;
     return Scaffold(
       body: SafeArea(
-          child: Column(
-        crossAxisAlignment: CrossAxisAlignment.center,
-        children: [
-          const SizedBox(
-            height: 10,
-          ),
-          const Text(
-            'Welcome to our Whatsapp',
-            style: TextStyle(fontSize: 33, fontWeight: FontWeight.w600),
-          ),
-          SizedBox(height: size.height / 9),
-          Image.asset(
-            'assets/bg.png',
-            height: 340,
-            width: 340,
-            color: tabColor,
-          ),
-          SizedBox(height: size.height / 15),
-          Padding(
-            padding: const EdgeInsets.all(15.0),
-            child: const Text(
-              'Read our Privacy Policy. Tap "Agree and Continue" to accept the Terms and Conditions ',
-              style: TextStyle(color: greyColor),
-              textAlign: TextAlign.center,
+          child: SingleChildScrollView(
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.center,
+          children: [
+            const SizedBox(
+              height: 10,
             ),
-          ),
-          const SizedBox(
-            height: 5,
-          ),
-          SizedBox(
-              width: size.width * 0.75,
-              child: CustomButton(
-                text: 'AGREE AND CONTINUE',
-                onPressed: () => navigateToLoginScreen(context),
-              )),
-        ],
+            const Text(
+              'Welcome to our Whatsapp',
+              style: TextStyle(fontSize: 33, fontWeight: FontWeight.w600),
+            ),
+            SizedBox(height: size.height / 12),
+            Image.asset(
+              'assets/bg.png',
+              height: size.height * 0.35 > 340 ? 340 : size.height * 0.35,
+              width: size.height * 0.35 > 340 ? 340 : size.height * 0.35,
+              color: tabColor,
+            ),
+            SizedBox(height: size.height / 20),
+            Padding(
+              padding: const EdgeInsets.all(15.0),
+              child: const Text(
+                'Read our Privacy Policy. Tap "Agree and Continue" to accept the Terms and Conditions ',
+                style: TextStyle(color: greyColor),
+                textAlign: TextAlign.center,
+              ),
+            ),
+            const SizedBox(
+              height: 5,
+            ),
+            SizedBox(
+                width: size.width * 0.75,
+                child: CustomButton(
+                  text: 'AGREE AND CONTINUE',
+                  onPressed: () => navigateToLoginScreen(context),
+                )),
+            const SizedBox(height: 20), // Bottom padding
+          ],
+        ),
       )),
     );
   }
