@@ -15,7 +15,7 @@ final selectContactControllerProvider = Provider((ref) {
 });
 
 class SelectContactController {
-  final ProviderRef ref;
+  final Ref ref;
   final SelectContactRepository selectContactRepository;
 
   SelectContactController(
